Avance Backend Demo Rimac

+++ backend-rimac-demo/src/main/java/com/rimac/demo/repository/InsuranceRepository.java
@@ -6,6 +6,6 @@ import com.rimac.demo.entity.Insurance;
 
 public interface InsuranceRepository extends CrudRepository<Insurance, Long> {
 
-	public Insurance findByInsuranceId(int insuranceId);
+	public Insurance findByInsuranceId(Long insuranceId);
 	
 }
